validação para o form de imóveis
Agora a validação dos campos das áreas depende do tipo de construção que é cadastrada. Também foi arrumado nas migrations para que as áreas fossem consideradas do tipo double e não integer, o que era um erro.

+++ database/migrations/2025_01_02_160410_create_imoveis_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('imoveis', function (Blueprint $table) {
             $table->id('inscricao_municipal');
             $table->string('tipo', 100);
-            $table->integer('area_terreno')->nullable();
-            $table->integer('area_edificacao')->nullable();
+            $table->double('area_terreno')->nullable();
+            $table->double('area_edificacao')->nullable();
             $table->string('logadouro', 100);
             $table->integer('numero');
             $table->string('bairro', 100);
